Add typings for account merge.
